tdf#115227 svtools: suppress UNO notifications for color selectors
This is nominally a regression from commit
3d1b4235fe41e7b6c55f2570ac845bebf429b86d (unify color selectors,
2016-10-28), but that just changed the Writer color picker to behave the
same way as the Impress one.

The Impress one started to emit these events with
ebcb381e5d347e25fed7fd876c7c287f6612561d (re-base on ALv2 code.
Includes:, 2012-11-15), to fix i#118707, improving accessibility.

That means either the focus changes and then accessibility is happy or
the focus does not change and then the UNO API client only gets events
when the user actually switches to an other window.

Fix the problem with suppressing UNO-level notifications for the
problematic events, by moving the existing VclListenerLock to a public
header and using it at two more places in svtools.

This should address not just color selectors, but in general other
toolbar menus / popup windows.

Change-Id: If427708c5b9fe4fa49cb8f00ec04b32cb28eb391
Reviewed-on: https://gerrit.libreoffice.org/48570
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>