tdf#115221 crash in printing and comment
not wonderfully happy with this fix, because we're changing the delete
ordering here, but the origin of this bug was

    commit c29f5c736a9a709a0601e63f336a352659b07d1c
    use rtl::Reference in SwDocFac instead of manual acquire/releas

and

    commit	cf7a168f37e4d4ae482122a9cdec37ba52b73c42 (patch)
    tdf#112292 - fix memory leak and use more auto ref counting in sw

so the delete ordering was obviously messed up in those already, and we
have such a morass of callbacks here, I can't see where else to fix it.

I tried to add a couple more "getDoc() != nullptr" checks, but we still
ended up crashing with an SwViewShell that had a bad pLayoutAccess
pointer.

Change-Id: I4aa39d58ae25bed41e3fd2c32c6d53659012ec10
Reviewed-on: https://gerrit.libreoffice.org/48658
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>