Correctly import solid fill color.
In orcus, a solid fill type uses the foreground color only and ignores
the background color.  Also, let's not use the alpha component as it
would cause the color to not get rendered at all.

Some patches are applied against liborcus in order to adjust the ODF
styles import code for this change.  These changes will be incorporated
in 0.13.3.

Change-Id: I9e8c243cc6a7f366de2393e7b7ecf77366f5f9ea
Reviewed-on: https://gerrit.libreoffice.org/49071
Reviewed-by: Kohei Yoshida <libreoffice@kohei.us>
Tested-by: Kohei Yoshida <libreoffice@kohei.us>