tdf#115218 framework: decouple HiddenForConversion from Hidden
Commit b03da8d7fcddaa85fe62ba3852445e81598962a0 (framework: disable
layout manager in hidden frames, 2018-01-17) added support for bypassing
the framework layout manager for doc conversion purposes, but the Hidden
flag has the meaning that the doc is only initially hidden, it may be
visible later, so reusing Hidden broke mail merge.

Introduce a dedicated HiddenForConversion flag, clients interested in
improved performance can opt in for that, and this way behavior for
Hidden is not changing.

Change-Id: If7537197900f8819f714b164c9e056b4ae69286d
Reviewed-on: https://gerrit.libreoffice.org/49302
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>