Include <release> in Flatpak appdata.xml
...fixing <https://github.com/flathub/org.libreoffice.LibreOffice/issues/13>
"No version information".  This is only minimal release information (cf.
<https://www.freedesktop.org/software/appstream/docs/chap-Metadata.html#
tag-releases>), but is what is available directly in the build without
additional manual input.  Lack of either a date or timestamp attribute causes a
failure "org.libreoffice.LibreOffice.desktop: AppData problem: attribute-missing
: <release> has no timestamp", so include the current date in ISO 8601 format to
silence that.

It is assumed that the LIBO_VERSION_* variables will only contain numeric
values, so don't need to be encoded when included in an XML attribute.

Change-Id: I442a01d5093ab2621897685c3bc1eeda4ee08fa9