Integrate initial version of gla11y tool in the build system
This is part of integrating an accessibility non-regression tool. This
adds checks in configure.ac for the presence of python lxml which we will
need, and adds support for calling the tool at build time, to check for
definite UI errors. For now, that only emits errors for missing or duplicate
accessibility relation targets, and senseless relations: a label being
mnemonic for several widgets.

Change-Id: Idda91b15b9a9e0322d16db33dfac8e03f2aa518c
Reviewed-on: https://gerrit.libreoffice.org/49856
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>