tdf#115873 sd navigator: adapt IsEqualToDoc() to Fill() further
SdPageObjsTLB::Fill() populates the navigator tree list box,
SdPageObjsTLB::IsEqualToDoc() determintes if it has to be refreshed or
not.

Commit bf91540b0b0fa694c08596a872d18e7e66339f61 (tdf#115873 sd
navigator: allow selecting but not focusing on objects, 2018-02-21)
already brought IsEqualToDoc() closer to Fill() with using the same
iteration mode for the SdrObjects, but that's not enough.

Fill() uses flat iteration, then checks for group shapes explicitly and
visits them recursively. Change IsEqualToDoc() to do the same, this way
selecting "Rectangle 3" in the testcase won't result in a jump back to
"Slide 1" (as an effect of IsEqualToDoc() returning false for an up to
date tree list box).

Change-Id: If2543cbc282af06ba43d4804e7ed455c8b9828cd
Reviewed-on: https://gerrit.libreoffice.org/50234
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>